backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
requests>=2.32.0
pydantic[email]>=2.9.0
python-dotenv>=1.0.1
sqlalchemy>=1.4.52
alembic>=1.13.1
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
pyjwt[crypto]>=2.10.0
pyotp>=2.9.0
gunicorn>=23.0.0
pydantic-settings>=2.7.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=2.10.0
email-validator>=2.2.0
prometheus-fastapi-instrumentator>=7.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
dnspython>=2.7.0
asyncssh>=2.18.0
